+++ src/main/java/org/betonquest/betonquest/conversation/SlowTellrawConvIO.java
@@ -6,6 +6,9 @@ import org.betonquest.betonquest.api.profiles.OnlineProfile;
 import org.betonquest.betonquest.utils.LocalChatPaginator;
 import org.betonquest.betonquest.utils.Utils;
 import org.bukkit.ChatColor;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.player.AsyncPlayerChatEvent;
 import org.bukkit.scheduler.BukkitRunnable;
 import org.jetbrains.annotations.Nullable;
 
@@ -21,6 +24,11 @@ public class SlowTellrawConvIO extends TellrawConvIO {
     @Nullable
     private List<String> endLines;
 
+    /**
+     * Whether the player can reply to the conversation, disabled while the NPC is talking, enabled when options are displayed
+     */
+    private boolean canReply;
+
     public SlowTellrawConvIO(final Conversation conv, final OnlineProfile onlineProfile) {
         super(conv, onlineProfile);
         final StringBuilder string = new StringBuilder();
@@ -34,6 +42,19 @@ public class SlowTellrawConvIO extends TellrawConvIO {
             delay = 10;
         }
         this.messageDelay = delay;
+        this.canReply = false;
+    }
+
+    /**
+     * if canReply is false, we ignore the event, otherwise handle it as normal
+     */
+    @EventHandler(priority = EventPriority.LOWEST, ignoreCancelled = true)
+    @Override
+    public void onReply(final AsyncPlayerChatEvent event) {
+        if (!canReply) {
+            return;
+        }
+        super.onReply(event);
     }
 
     @Override
@@ -42,6 +63,7 @@ public class SlowTellrawConvIO extends TellrawConvIO {
             end();
             return;
         }
+        canReply = false;
 
         // NPC Text
         final String[] lines = LocalChatPaginator.wordWrap(
@@ -63,6 +85,7 @@ public class SlowTellrawConvIO extends TellrawConvIO {
                     }
 
                     endLines = null;
+                    canReply = true;
 
                     this.cancel();
                     return;